move new test cases to test_textinput
