+++ testbed/tests/widgets/test_textinput.py
@@ -4,6 +4,8 @@ import pytest
 
 import toga
 from toga.constants import CENTER
+from toga.style import Pack
+from toga.style.pack import RIGHT, SERIF
 
 from ..data import TEXTS
 from .conftest import build_cleanup_test
@@ -247,3 +249,34 @@ async def test_undo_redo(widget, probe):
     await probe.redraw(f"Widget value should be {text_1!r}")
     assert widget.value == text_1
     assert probe.value == text_1
+
+
+async def test_no_event_on_initialization(widget, probe, on_change):
+    "The widget don't fire events on initialization."
+    # When the widget is created and added to a box, no on_change event is fired.
+    parent = toga.Box(style=Pack(flex=1))
+    parent.add(widget)
+    await probe.redraw("Widget should not fire events on initialization")
+    on_change.assert_not_called()
+    on_change.reset_mock()
+
+
+async def test_no_event_on_style_change(widget, probe, on_change):
+    "The widget don't fire on_change events on text style changes."
+    # font changes
+    widget.style.font_family = SERIF
+    await probe.redraw("Font style has been changed")
+    on_change.assert_not_called()
+    on_change.reset_mock()
+
+    # text alignment changes
+    widget.style.text_align = RIGHT
+    await probe.redraw("Text alignment has been changed")
+    on_change.assert_not_called()
+    on_change.reset_mock()
+
+    # text color changes
+    widget.style.color = "#0000FF"
+    await probe.redraw("Text color has been changed")
+    on_change.assert_not_called()
+    on_change.reset_mock()
